Update french-apa.lbx
Proposed translation for most francophone values in \DeclareBibliographyStrings.
I have doubts about certain translations. The fields 'tothe', 'ofthe' and 'present' remain obscure to me, as I cannot see their context of use, and the correct translation obviously depends on this context. 
For the 'powerpoint' field value, I have retained the (commercial) application name, but this is probably not the best choice, as it does not account for the variety of formats (odp, pdf...). Usage in French is not standardised, but one might prefer: powerpoint = {{presentation\space slides}{slides}},
For 'archivedat', I feminised the values, as the term 'data' is feminine in French.
I modified the values for 'chair' and 'chairs', because in French, in most francophone countries, a 'Maître de conférences' is a lecturer (tenured in France and French-speaking Africa, professional course leader in Belgium and Switzerland).
Regarding the fields 'typevolseries', 'typevolseriess', I question this: I have never encountered a reference indicating this double information. In most cases, one simply uses 'dir.' for all types of responsibilities (except translator: 'trad.').